add champion triche

+++ corewar/src/in_fork.c
@@ -6,7 +6,7 @@
 /*   By: sle-lieg <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/12 16:36:38 by sle-lieg          #+#    #+#             */
-/*   Updated: 2017/05/15 14:35:19 by folkowic         ###   ########.fr       */
+/*   Updated: 2017/05/17 20:51:47 by folkowic         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void in_fork(t_player *player)
 	int			value;
 
 	curs = (player->pc + 1) % MEM_SIZE;
-	value = vm_get_param_val(curs, 2) % IDX_MOD;
+	value = (short)vm_get_param_val(curs, 2) % IDX_MOD;
 	vm_lst_add_cpy(&player);
 	player->pc = (player->pc + 3) % MEM_SIZE;
 	player->next->cycles_cd++;
